requirements.txt
streamlit
sentence-transformers
PyMuPDF
faiss-cpu
selenium
langchain
langchain-community
huggingface-hub
python-dotenv
